python/src/auth/requirements.txt
Flask==3.1.3
PyMySQL==1.1.1
PyJWT==2.11.0
cryptography==44.0.2
